So it turns out all I needed to do to get Planisphere rendering properly... was update the internal render structures to accept the 32-bit seg indicies that DeepBSP builds. Derp.

+++ src/doom/r_defs.h
@@ -285,8 +285,8 @@ typedef struct line_s
 typedef struct subsector_s
 {
 	sector_t*			sector;
-	uint16_t			numlines;
-	uint16_t			firstline;
+	uint32_t			numlines;
+	uint32_t			firstline;
 	int32_t				index;
 } subsector_t;
